// types.ts
export interface CosmicObject {
  id: string
  slug: string
  title: string
  content?: string
  metadata: Record<string, unknown>
  type: string
  created_at: string
  modified_at: string
}

export interface CosmicImage {
  url: string
  imgix_url: string
}

export interface Category extends CosmicObject {
  type: 'categories'
  metadata: {
    name?: string
    description?: string
  }
}

export interface Author extends CosmicObject {
  type: 'authors'
  metadata: {
    name?: string
    bio?: string
    avatar?: CosmicImage
  }
}

export interface Post extends CosmicObject {
  type: 'posts'
  metadata: {
    content?: string
    featured_image?: CosmicImage
    author?: Author
    category?: Category
    location?: string | null
  }
}